feat: 共通化 & 抽象することによって依存クラスが少なくなるようにした
