+++ modules/data/src/main/java/net/pantasystem/milktea/data/di/module/UserModule.kt
@@ -72,4 +72,7 @@ abstract class UserModule {
     @Singleton
     internal abstract fun bindUserApiAdapter(impl: UserApiAdapterImpl): UserApiAdapter
 
+    @Binds
+    @Singleton
+    internal abstract fun bindUserCacheUpdater(impl: UserCacheUpdaterFromUserActionResultImpl): UserCacheUpdaterFromUserActionResult
 }
