+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/user/mute/MuteRepositoryImpl.kt
@@ -5,19 +5,15 @@ import kotlinx.coroutines.withContext
 import net.pantasystem.milktea.common.Logger
 import net.pantasystem.milktea.common.runCancellableCatching
 import net.pantasystem.milktea.common_android.hilt.IODispatcher
-import net.pantasystem.milktea.data.infrastructure.toUserRelated
-import net.pantasystem.milktea.data.infrastructure.user.UserActionResult
+import net.pantasystem.milktea.data.infrastructure.user.UserCacheUpdaterFromUserActionResult
 import net.pantasystem.milktea.model.user.User
-import net.pantasystem.milktea.model.user.UserDataSource
-import net.pantasystem.milktea.model.user.UserRepository
 import net.pantasystem.milktea.model.user.mute.CreateMute
 import net.pantasystem.milktea.model.user.mute.MuteRepository
 import javax.inject.Inject
 
 internal class MuteRepositoryImpl @Inject constructor(
     private val muteApiAdapter: MuteApiAdapter,
-    private val userDataSource: UserDataSource,
-    private val userRepository: UserRepository,
+    private val updateCacheFrom: UserCacheUpdaterFromUserActionResult,
     @IODispatcher private val coroutineDispatcher: CoroutineDispatcher,
     loggerFactory: Logger.Factory,
 ) : MuteRepository {
@@ -58,18 +54,5 @@ internal class MuteRepositoryImpl @Inject constructor(
         }
     }
 
-    private suspend fun updateCacheFrom(userId: User.Id, result: UserActionResult, reducer: suspend (User.Detail) -> User.Detail) {
-        val user = userRepository.find(userId, true) as User.Detail
-        val updated = when(result) {
-            is UserActionResult.Mastodon -> {
-                user.copy(
-                    related = result.relationship.toUserRelated()
-                )
-            }
-            UserActionResult.Misskey -> {
-                reducer(user)
-            }
-        }
-        userDataSource.add(updated)
-    }
+
 }
